Tron-Backend/requirements.txt
fastapi
uvicorn
motor[srv]
beanie
python-dotenv
pydantic[email]
logging
tronpy